+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 - Removed `xrange()` method of the `Period` class and made `range()` a generator.
 - New locale system which uses CLDR data for most of the translations.
 - `diff_for_humans()` now returns `a few seconds` where appropriate.
+- Removed `Period.intersect()`.